+++ src/config/db.js
@@ -1,1 +1,8 @@
-module.exports = require('./').db;
+// @flow
+
+// Used for sequelize-cli migrations.
+
+require('babel-core/register');
+require('../base');
+
+module.exports = require('./').default.db;