+++ src/config/dev.js
@@ -1,14 +1,28 @@
-module.exports = {
+// @flow
+
+import { minutes } from 'lib/time';
+
+export default {
   db: {
     host: process.env.DB_PORT_3306_TCP_ADDR,
   },
 
   leaderboards: {
-    refreshInterval: 60000 * 1,
-    getCacheTtl: 1,
+    latestSeasonCacheTtl: minutes(1),
+    refreshInterval: minutes(1),
+    getCacheTtl: minutes(1),
   },
 
   fetch: {
-    interval: 60000 * 1,
+    interval: minutes(1),
+  },
+
+  cache: {
+    findAllCharacters: minutes(1),
+    statistics: minutes(1),
+  },
+
+  web: {
+    publicUrl: 'http://localhost:3000',
   },
 };