+++ src/config/local.js
@@ -1,4 +1,6 @@
-module.exports = {
+// @flow
+
+export default {
   db: {
     // host: '127.0.0.1',
     host: process.env.DB_PORT_3306_TCP_ADDR,