+++ src/config/test.js
@@ -1,9 +1,13 @@
-module.exports = {
+// @flow
+
+import { minutes } from 'lib/time';
+
+export default {
   db: {
     host: 'gw2armory-test.cekbcmynaoxp.us-east-1.rds.amazonaws.com',
   },
 
   fetch: {
-    interval: 60000 * 1,
+    interval: minutes(1),
   },
 };