+++ src/lib/time.js
@@ -1,0 +1,9 @@
+// @flow
+
+export function minutes (n: number) {
+  return 60000 * n;
+}
+
+export function hours (n: number) {
+  return 60000 * 60 * n;
+}